[versig] [WKI90666] Ness 4.1.4: Alter savupdate to download the extra files.
Alter versig to allow time hack to be turned off.
Alter extra files downloading to download manifest.dat and use that as the index.
Add a new message for failing to verify a manifest before we download files.
Add a test to check we download with the manifest.dat not the index.spec.
Alter create_certificates.sh to not include text output in the signing certificate.

[+OST]

[git-p4: depot-paths = "//dev/savlinux/trunk/products/versig/": change = 2319897]