[base] [LINUXEP-6013] Add Proxy implementation.
Allows Pub-Sub architecture with both Pubs and Subs connecting to a central Databroker.
ProxyImpl has a back-door to get the ContextHolder out, so that we can use inproc:// sockets for unit tests.

Move the timeout setup code into SocketUtil to allow sharing.
Move listen code to SocketUtil to allow sharing.
Add a reset() method to ContextHolder to allow the context to be deleted before the object is destroyed to support ProxyImpl usage.

Switch ZeroMQWrapperImpl unit tests to a single executable to match preferred usage of googletest.