[base] [LINUXEP-6511] Check remove works or the file doesn't exist.
Fixes:
31502 Unchecked return value from library

The function returns a value that indicates an error condition. If this is not checked, the error condition may not be handled correctly.

In <unnamed>::​FileSystemImplTest_atomicWriteStoresExpectedContentForFile_Test::​TestBody(): Value returned from a library function is not checked for errors before being used. This value may indicate an error condition. (CWE-252)