[base] [LINUXEP-6511] Check return codes from fcntl calls.
Make against pipes they shouldn't ever fail.

Fixes:
31501 Unchecked return value from library

The function returns a value that indicates an error condition. If this is not checked, the error condition may not be handled correctly.

In <unnamed>::​setNonBlocking(int): Value returned from a library function is not checked for errors before being used. This value may indicate an error condition. (CWE-252)