[base] [LINUXEP-6511] Isolate mkdir and check its return code.
Fixes:
31492 Unchecked return value from library

The function returns a value that indicates an error condition. If this is not checked, the error condition may not be handled correctly.

In <unnamed>::​FileSystemImplTest_readFileUsingDirectoryPathShouldThrow_Test::​TestBody(): Value returned from a library function is not checked for errors before being used. This value may indicate an error condition. (CWE-252)