[base] [LINUXEP-6511] Check ::remove return code.
Fixes:
31486 Unchecked return value from library

The function returns a value that indicates an error condition. If this is not checked, the error condition may not be handled correctly.

In RealiNotifyDirectoryWatcherTests_DeleteFilesDoesNotTriggerEvent_Test::​TestBody(): Value returned from a library function is not checked for errors before being used. This value may indicate an error condition. (CWE-252)