[base] Replace hash map approach to converting string -> command with a more basic if {} else if{} block to avoid data race.
If two threads entered mapStringCommand, the first count start adding entries (so .empty() is false), then the second could enter and try and resolve an entry, then the first continue adding entries (invalidating iterators).

I don't think this is the cause of the segfaults but it was an error detected by helgrind in our code.