[base] [LINUXEP-7053] Remove setAppIds from the interface since the code isn't using it.
register is only called with a direct object.
Pass the lock into the lock functions to force the lock to exist - makes it harder to call incorrectly.
Make registerPlugin
Change the comments - we always create a new plugin, so that we reconnect to the 0MQ socket.
Fix up ManagementAgentMain and unit tests to use the new function, which has been added to the interface.