[base] [LINUXEP-7266] Make the m_stopThread atomic
m_stopThread of SenderThread is a normal boolean variable, but it is accessed across different threads. This is a race condition and is undefined behaviour. This could explain the segfault, but could also be in zmq or somewhere else.
Just an attempt.
atomics can be used in multithread without being a race condition.