[base] [Coverity] Avoid throwing from getInstance.
(It's called from destructors, and log methods shouldn't throw).

33973 Uncaught exception
If the exception is ever thrown, the program will crash.
In Common::​PluginApiImpl::​SensorDataSubscriber::​~SensorDataSubscriber(): A C++ exception is thrown but never caught (CWE-248)

33974 Uncaught exception
If the exception is ever thrown, the program will crash.
In watchdog::​watchdogimpl::​PluginProxy::​~PluginProxy(): A C++ exception is thrown but never caught (CWE-248)