[base] [LINUXEP-7309] Add Subscription and Features to ConfigurationData
In order to minimise the impact in SulDownloader and UpdateScheduler and the SystemTests, the
ConfigurationData was introduced in 'compatible with previous version'. It will not need to be kept
after the systemtests, suldownloader and updatescheduler are updated to the new version.

The only thing that would need to be discussed is support for upgrade.
A suggestion is made that if SulDownloader is not able to understand a configuration file it should
remove if to allow UpdateScheduler to recreate it.