[base] [LINUXEP-8233] Fix Suldownloader to re-install plugin if necessary
When a product is uninstalled, the install.sh is removed from the local primary distributed path. This allows that the product will be recognized as changed if it is downloaded afterwards.
Another possibility would be to remove the whole folder ensuring that the files and libraries are completely removed from disk on uninstallation.

Update tests.