[base] [LINUXEP-8190] Add first Lib Fuzz test target for ManagementAgent
LibFuzz only works with clang. For this reason, some 'tricks' were added to make
possible for developers to develop fuzz targets and to run it not on developers machine.

CMakeLists: handle the different compiler options when building Fuzz tests.
tests/CMakeLists: add LibFuzzerTests as subdirectory
tests/buildFuzzTests.sh: script that get all that is necessary to build target ManagementAgentApiTest.
tests/CMakeLists.txt: setup of ManagementAgentApiTest
tests/message.proto: Definition of the protobuf message which is the input of the target.
tests/ManagementAgentApiTest.cpp: Description of the test.