[base] [LINUXDAR-695] Revert debug mode as it breaks pluginapi
Plugins are now failing to compile because of the debug build of pluginapi.

As is now require asan.
14:13:03  stdout: [ 75%] Linking CXX executable mtr
14:13:04  stdout: ../../../redist/pluginapi/lib64/libpluginapi.so: undefined reference to `__asan_report_store4'
14:13:04  stdout: ../../../redist/pluginapi/lib64/libpluginapi.so: undefined reference to `__asan_stack_free_7'

There are two possible fixes:
1. Add the sanitize-address to the plugins when they use filer6 outputs (debug outputs) (my preference)
2. Make the pluginapi not to have the sanitizers on, hence, not causing the issue with the plugins.

As this is not in the scope for this sprint, I suggest we revert for the moment. Keeping how to add the correct flags in CI
and have this sorted afterwards.