[base] [LINUXDAR-756] Changed telemetry to have a 'stick bit'
Handles the case where we want telemetry to always send the information,
even if there was no update since the last call to telemetry.

This avoids the need to increment(0)
The code is prepeared to change the increment(0) but it is not in the scope of this ticket.
Avoid the need to use the callback for telemetry.