[base] [LINUXDAR-882] Add time to watchdog to start all managed services
Recent failure of test happened because a plugin was sent a SIGTERM(15) signal before it had even time to setup the handlers.
This was due to the startSsplService instantiate the watchdog but gives it not enough time to start all the managed services.
Adding a small time should suffice to avoid this problem

Extract of the failure below:

[2019-12-13T05:46:14.148]    INFO ProcessMonitoringImpl <> Starting /opt/sophos-spl/plugins/mtr/bin/mtr
[2019-12-13T05:46:14.170]    INFO watchdog <> Command from IPC: STOP
[2019-12-13T05:46:14.170]    INFO watchdog <> Requesting stop of mtr
[2019-12-13T05:46:14.170]    WARN ProcessMonitoringImpl <> /opt/sophos-spl/plugins/mtr/bin/mtr still running, despite being disabled: 3
[2019-12-13T05:46:14.170]    INFO ProcessMonitoringImpl <> Stopping /opt/sophos-spl/plugins/mtr/bin/mtr
[2019-12-13T05:46:14.194]    INFO ProcessMonitoringImpl <> Output:
[2019-12-13T05:46:14.194]   ERROR ProcessMonitoringImpl <> /opt/sophos-spl/plugins/mtr/bin/mtr died with 15