[av] [LINUXDAR-1436] Read and deserialise the Capn Proto config file
Moved the .capnp definition to "scan_messages" so that it can be more readily shared with the avscanner.
Created an additional constructor for NamedScanRunner for convenient unit-testing without creating files.
Unit test currently only checks the scan name need to extend.
Perhaps eventually remove the getter methods and test by seeing the exclusions passed on to a Mock filewalker.