[av] [LINUXDAR-1428] Pass archive setting to scanner.
We need to pass archive setting to the scanner, so that we can configure our threat engine to scan inside archives or not.
Add interface to ScanningClientSocket to allow easier testing of ScanClient.
Add a ScanningClientSocket scan method that takes a request, to avoid having to pass other arguments across the interface.