[base] [LINUXDAR-1480] Remove the hack to forward authorization to the push server as it is not necessary
The authorization needs to be sent to the load balancer that will redirect to the push server.
But it does not need to go to the push server. The load balancer will provide a redirect location
that handles the endpoint authorization.

This means that the RequestsSessionAllowRedicrectWithAuth can be removed.

Handle adjustments to tests.