[base] [LINUXDAR-1924] prevent the managementagent running in the fuzzer from interacting with the telemetry on shutdown
The fuzzer runs in two threads and is using two static objects (telemetry) and the (managementagent). The order that they are destructed is causing seg-fault. As this is not relevant to the test in case, hence, added the capability of avoiding the usage of the save method on the telemetry for persistency