[base] [LINUXDAR-1649] [LINUXDAR-1958] [LINUXDAR-1942] Remove SubscriptionComponent from DownloadReport
The bug reported in LINUXDAR-1958 means that the DownloadReport can not have any extra field or downgrade would fail.
This commit removes the extra field proposed (SubscriptionComponent) and re-implement it by means of reporting in the ProductStatusReport which is the one that old versions of the product will use to calculate the alc status. This now fix the problem of downgrade. But, it would be nice for in future go back to the easier to understand solution
The new ProductStatusReport is calculated by combining the information about the subscriptions and the downloaded product. If a subscription involves a group of downloaded products, the error and status of the subscription is calculated in relation to the products involved