[av] [LINUXDAR-2176] Remove default argument value from virtual function.
Clang-tidy warns: Warning:(22, 22) Clang-Tidy: Default arguments on virtual or override methods are prohibited

See https://stackoverflow.com/questions/27974653/default-parameter-in-virtual-functions-c

The problem with specifying a default argument for symlinkTarget in the C++ is that any variable pointing to a concrete sub-class wouldn't have it, or could have a different value.

Instead we provide a non-virtual inline function that provides the default argument directly.