[av] [LINUXDAR-2999][LINUXDAR-2996] Careful examination shows that the init lock needs to be help in update(), while examining if init has not happened and saving the update details if so.
The failure case would be:

THREAD1: update() -> as far as m_updatePath, then suspended
THREAD2: init() -> completion
    won't update sincec m_updatePending not set
THREAD1: Resumes, won't update, since already in if block.

Result: no update happens despite update() being called.

In practice update() is called during start up, so this sequence is unlikely to ever happen, but better to be safe, and update() is a rare function to be called.