[av] [CORE-2797] Update engine to 3.82.2
Hi Jason,

It’s already available

We have produced a release candidate of 3.82.2 engine with a fix for CORE-2797 Crash in 3.82.1 engine
https://artifactory.sophos-ops.com/esg-build-candidate/core.engine/release--3.82.1/20210706101701-372b5c3d0db8684173bd959152d13fd2c4e2a7e3-cO07yp/build/

Although the path says “3.82.1”, it is a 3.82.2 engine (we did it to make it easier for out automation to pick up a new version)

The test is in progress, but we have performed a smoke test of scanning and detecting eicar, which worked ok.
If rest of the testing goes well, I’m hoping to have a release review by mid of next week.

With kind regards
mario