[sdds] [EES-21716] Correct furcated entitlement.
Hi Douglas Leeder,

I think the problem is you've not told Ballista the correct furcated entitlement of your warehouse so it doesn't know which warehouse it should replace.

Your pubspec says

<warehouse furcatedentitlement="SSPL" id="sspl">
but I think it should say

<warehouse furcatedentitlement="LINUX_SSPL" id="sspl">
The furcated entitlement of the SSPL warehouse in Cred Admin is 'LINUX_SSPL'.