[av] [LINUXDAR-3206] - Add threatMemoryUsage and threatProcessAge to getTelemetry in PluginCallback by accessing the proc/pid/stat file.
Re-factor code out of CalculateHealth to a separate function for getting pid

Adjustments to existing unit tests to ensure they dont return false negatives. Add a number of unit tests related to new functionality. This necessitated the addition of a wrapper for the call to sysInfo so it could be mocked.

Add a single test to AVAndBase.robot to ensure that a call to getTelemetry is returning non-zero