[av] Reduce number of eicars in "CLS Summary is Correct" test.
For AV develop build:
https://portal.sophos-ops.com/build-stats/20230227153439-d52d3345e84ffdfb27c20208edbc3fc0a48c860d-X6OvOr
develop#884
(and previously)

We had a failure of "CLS Summary is Correct" - Global Reputation timeout, causing avscanner to report 2 scan errors, and tear down detects errors in threat detector.

As a short-term fix, reduce the number of eicars; since every eicar requires a GR lookup for false-positives.

In the long-term we might want to run this test without LiveProtection, or alter error reporting to ignore GR timeouts?