[TESTFIX] Fix AV SUSI initialisation failures (#574)
One test failed because SUSI wasn't initialised before it attempted an update.
The other test fails because it checks for SUSI Loaded old data which wasn't logged on initialisation, only on updates. The test can sometimes cause test data to be loaded during initialisation rather than update, but neither is wrong for the purpose of the test.